Exercise stack trace rendering through the real API
