+++ cwms-data-api/src/test/java/cwms/cda/api/TimeSeriesFilteredControllerTestIT.java
@@ -3,12 +3,9 @@ package cwms.cda.api;
 import static io.restassured.RestAssured.given;
 import static io.restassured.config.JsonConfig.jsonConfig;
 import static org.hamcrest.Matchers.closeTo;
-import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
-import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -16,7 +13,6 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.flogger.FluentLogger;
 import cwms.cda.formatters.Formats;
-import fixtures.StackTraceFeatureExtension;
 import fixtures.TestAccounts;
 import io.restassured.RestAssured;
 import io.restassured.filter.log.LogDetail;
@@ -29,10 +25,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.junit.jupiter.api.Test;
 
 @Tag("integration")
 class TimeSeriesFilteredControllerTestIT extends DataApiTestIT {
@@ -458,59 +452,6 @@ class TimeSeriesFilteredControllerTestIT extends DataApiTestIT {
         } catch (SQLException ex) {
             throw new RuntimeException("Unable to create location for TS", ex);
         }
-    }
-
-    @Test
-    @ExtendWith(StackTraceFeatureExtension.class)
-    void returnsStackTraceLinesForTraceRoleWhenRequestFails() throws Exception {
-        ObjectMapper mapper = new ObjectMapper();
-
-        InputStream resource = this.getClass().getResourceAsStream(JSON_FILE);
-        assertNotNull(resource);
-        String tsData = IOUtils.toString(resource, StandardCharsets.UTF_8);
-
-        JsonNode ts = mapper.readTree(tsData);
-        String location = ts.get(Controllers.NAME).asText().split("\\.")[0];
-        String officeId = ts.get("office-id").asText();
-
-        createLocation(location, true, officeId);
-
-        TestAccounts.KeyUser user = TestAccounts.KeyUser.SPK_NORMAL2;
-
-        given()
-            .log().ifValidationFails(LogDetail.ALL, true)
-            .accept(Formats.JSONV2)
-            .contentType(Formats.JSONV2)
-            .body(tsData)
-            .header("Authorization", user.toHeaderValue())
-            .queryParam(Controllers.OFFICE, officeId)
-        .when()
-            .redirects().follow(true)
-            .redirects().max(3)
-            .post("/timeseries/")
-        .then()
-            .log().ifValidationFails(LogDetail.ALL, true)
-            .assertThat()
-            .statusCode(is(HttpServletResponse.SC_OK));
-
-        given()
-            .log().ifValidationFails(LogDetail.ALL, true)
-            .accept(Formats.JSONV2)
-            .cookie("JSESSIONIDSSO", user.getJSessionId())
-            .queryParam(Controllers.OFFICE, officeId)
-            .queryParam(Controllers.NAME, ts.get(Controllers.NAME).asText())
-            .queryParam(Controllers.BEGIN, "not-a-date")
-        .when()
-            .redirects().follow(true)
-            .redirects().max(3)
-            .get("/timeseries/filtered/")
-        .then()
-            .log().ifValidationFails(LogDetail.ALL, true)
-            .assertThat()
-            .statusCode(is(HttpServletResponse.SC_BAD_REQUEST))
-            .body("incidentIdentifier", notNullValue())
-            .body("details.stackTraceLines.size()", greaterThan(0))
-            .body("details.stackTraceLines[0]", containsString("Exception"));
     }
 
 }
